Move internal implementations in separate package
