+++ plugin/src/test/groovy/com/novoda/buildproperties/internal/MapEntriesTest.groovy
@@ -1,5 +1,7 @@
-package com.novoda.buildproperties
+package com.novoda.buildproperties.internal
 
+import com.novoda.buildproperties.Entry
+import com.novoda.buildproperties.internal.MapEntries
 import org.junit.Test
 
 import static com.novoda.buildproperties.test.ExtendedTruth.assertThat
